Rename chart folder
