+++ src/components/UI/Graphics/Chart.interface.ts
